Revert "Feature/client"
