+++ client/src/store/types.ts
@@ -1,0 +1,16 @@
+export interface Product {
+  B: boolean
+  C: number
+  CV: null
+  G: number
+  P: number
+  Pl: null
+  T: number
+}
+
+export interface PreparedProduct {
+  id: number
+  groupId: number
+  price: number
+  inStock: number
+}
